% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cuasm_insns}
\alias{cuasm_insns}
\title{Get decoded SASS instructions, optionally filtered to one kernel.}
\usage{
cuasm_insns(obj, kernel = NULL)
}
\arguments{
\item{obj}{A \code{cuasm} object.}

\item{kernel}{Optional kernel name (string).}
}
\value{
data.frame with kernel, slot, address, text, instr_hex, ctrl_hex.
}
\description{
Get decoded SASS instructions, optionally filtered to one kernel.
}
