% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cuasm_kernels}
\alias{cuasm_kernels}
\title{List the kernel \code{.text} sections in a cubin.}
\usage{
cuasm_kernels(obj)
}
\arguments{
\item{obj}{A \code{cuasm} object.}
}
\value{
data.frame with kernel, offset (file offset), size.
}
\description{
List the kernel \code{.text} sections in a cubin.
}
