% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cuasm_save_cuasm}
\alias{cuasm_save_cuasm}
\title{Save a human-readable .cuasm-style text dump.}
\usage{
cuasm_save_cuasm(obj, out_path)
}
\arguments{
\item{obj}{A \code{cuasm} object.}

\item{out_path}{Destination path (typically \code{.cuasm}).}
}
\value{
out_path (invisibly).
}
\description{
Save a human-readable .cuasm-style text dump.
}
