% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cuasm_sections}
\alias{cuasm_sections}
\title{Show all ELF sections in a cuasm object.}
\usage{
cuasm_sections(obj)
}
\arguments{
\item{obj}{A \code{cuasm} object.}
}
\value{
data.frame with index, name, type, offset, size.
}
\description{
Show all ELF sections in a cuasm object.
}
